+++ accounts-service/src/main/java/ru/yandex/practicum/accounts/exception/InsufficientFundsException.java
@@ -2,14 +2,14 @@ package ru.yandex.practicum.accounts.exception;
 
 public class InsufficientFundsException extends RuntimeException {
 
-    private final Integer availableBalance;
+    private final Long availableBalance;
 
-    public InsufficientFundsException(Integer availableBalance) {
+    public InsufficientFundsException(Long availableBalance) {
         super("Insufficient funds. Available: " + availableBalance);
         this.availableBalance = availableBalance;
     }
 
-    public Integer getAvailableBalance() {
+    public Long getAvailableBalance() {
         return availableBalance;
     }
 }
