+++ accounts-service/src/test/java/ru/yandex/practicum/accounts/util/TestDataFactory.java
@@ -7,7 +7,7 @@ import java.time.LocalDate;
 import java.util.List;
 
 public class TestDataFactory {
-    public static Account createTestAccount(String login, String name, LocalDate birthdate, Integer balance) {
+    public static Account createTestAccount(String login, String name, LocalDate birthdate, Long balance) {
 
         Account account = new Account();
         account.setLogin(login);
@@ -18,15 +18,15 @@ public class TestDataFactory {
     }
 
     public static Account createIvanovAccount() {
-        return createTestAccount("ivanov", "Иван Иванов", LocalDate.of(1990, 1, 15), 5000);
+        return createTestAccount("ivanov", "Иван Иванов", LocalDate.of(1990, 1, 15), 5000L);
     }
 
     public static Account createPetrovAccount() {
-        return createTestAccount("petrov", "Петр Петров", LocalDate.of(1985, 5, 20), 3000);
+        return createTestAccount("petrov", "Петр Петров", LocalDate.of(1985, 5, 20), 3000L);
     }
 
     public static Account createSidorovAccount() {
-        return createTestAccount("sidorov", "Сидор Сидоров", LocalDate.of(1995, 11, 30), 1000);
+        return createTestAccount("sidorov", "Сидор Сидоров", LocalDate.of(1995, 11, 30), 1000L);
     }
 
     public static UpdateAccountRequest createUpdateAccountRequest(String name, LocalDate birthdate) {
