+++ cash-service/src/main/java/ru/yandex/practicum/cash/dto/CashOperationRequest.java
@@ -6,7 +6,7 @@ import jakarta.validation.constraints.Positive;
 public record CashOperationRequest(
         @NotNull(message = "Value is required")
         @Positive(message = "Value must be positive")
-        Integer value,
+        Long value,
         @NotNull(message = "Action is required")
         CashAction action
 ) {
