+++ transfer-service/src/main/java/ru/yandex/practicum/transfer/dto/TransferResponse.java
@@ -3,7 +3,7 @@ package ru.yandex.practicum.transfer.dto;
 public record TransferResponse(
         boolean success,
         String message,
-        Integer senderBalance,
-        Integer recipientBalance
+        Long senderBalance,
+        Long recipientBalance
 ) {
 }
